Final Project/requirements.txt
numpy==1.24.4
opencv_python==4.10.0.84
pyttsx3==2.98
PyYAML==6.0.2
torch==2.4.1
tqdm==4.67.0
ultralytics==8.3.32
